+++ HISTORY.md
@@ -1,5 +1,9 @@
 # MediaWiki-Phan-Config release history #
 
+## UNRELEASED
+* Add plugin to disallow use of `new Exception` (Daimona Eaytoy)
+* Do not emit MediaWikiNoEmptyIfDefined for properties of classes with the AllowDynamicProperties attribute (Daimona Eaytoy) 
+
 ## 0.13.0 / 2023-09-08
 * Add plugin to forbid `empty()` on defined variables and properties (Daimona Eaytoy)
 * Bump phan to 5.4.2 and taint-check to 5.0.0 (Michael Große)